pyproject.toml
[tool.poetry]
name = "tap-mssql"
version = "2.6.0"
description = "A pipelinewise compatible tap for connecting Microsoft SQL Server"
authors = ["Rob Winters <wintersrd@gmail.com>"]
license = "GNU Affero"
readme = "README.md"
homepage = "https://robertwinters.nl"
repository = "https://github.com/wintersrd/pipelinewise-tap-mssql"
keywords = ["singer", "meltano", "pipelinewise", "mssql"]
classifiers = ["Development Status :: 5 - Production/Stable", "Topic :: Database"]

[tool.poetry.scripts]
tap-mssql = 'tap_mssql.__init__:main'

[tool.poetry.dependencies]
python = ">=3.8,<3.13"
attrs = ">=24.2.0"
pendulum = ">=1.2.0"
realit-singer-python = ">=5.0.0"
# pymssql==2.2.8 broken: https://github.com/pymssql/pymssql/issues/833
pymssql = ">=2.1.4,!=2.2.8"
backoff = ">=1.8.0"

[tool.poetry.dev-dependencies]
docker = "^7.1.0"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
mypy = "^1.11.1"
pytest-mock = "^3.14.0"
pytest-sugar = "^1.0.0"
pytest-datafiles = "^3.0"
tox-docker = "5.0.0"
pre-commit = "^3.5.0"
bumpversion = "^0.6.0"
requests = "2.32.3"
tox = "^4.18.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
